requirements.txt
crewai
crewai[tools]
python-dotenv
langchain
langchain-community
langchain-openai
langchain-huggingface